3-naive-rag/requirements.txt
langchain==1.2.17
langchain-community
langchain-text-splitters==1.1.2
langchain-openai==1.2.1
langchain-qdrant==1.1.0
qdrant-client==1.17.1
python-dotenv
pypdf==6.10.2
